Do not disable SSH agent
The current logic around the SSH agent is flawed and tentatives to fix
it to cover more cases only leads to more complexity.

As seen in #648, to this day, nobody really knows why we disable the
agent. In the meantime, many valid use cases just don't work because of
disabling the agent, including the good practice to use encrypted
private key files. I think it is time to bite the bullet and just leave
the agent always on: "ssh" from OpenSSH does not have an option to
disable the agent and nobody complains about this.